Project template completed
