+++ css/common.css
@@ -110,6 +110,6 @@ a {
   height: 100vh;
   width: 100vw;
   position: absolute;
-  background: #00000060;
-  z-index: 2;
+  background: #0000006c;
+  z-index: 1;
 }
