+++ index.html
@@ -53,12 +53,12 @@
         <section class="headline-container">
             <article class="headline">
                 <h2>"Hello! Sharing world"</h2>
-                <h2>CREATIVE COMMONS<br>GLOBAL SUBMMIT<br>2015</h2>
+                <h2>CREATIVE COMMONS<br>GLOBAL SUBMMIT<br>2023</h2>
                 <p class="intro-txt">
                     A joyfull celebration believing in the value of openness and sharing creating 
                     a positive change with people from over 80 countries is atking palce in Octobar, in Korea
                 </p>
-                <h2 class="time">2015.10.15(THU) ~ 16(FRI)</h2>
+                <h2 class="time">2023.10.15(THU) ~ 16(FRI)</h2>
                 <p>@ National Museum of Korea, Art Center Nabi and more</p>
             </article>
         </section>
@@ -96,7 +96,7 @@
             </article>
         </section>
         <h2 class="Program-foot dk">SEE THE WHOLE PROGRAM</h2>
-        <h2 class="Program-foot mb">Join CC Global Summit 2015</h2>
+        <h2 class="Program-foot mb">Join CC Global Summit 2023</h2>
     </section>
     <section id="speakers">
 
@@ -119,7 +119,7 @@
         <section class="buttom-f-nav">
             <a class="logo logo-abt" href="#">BLK-NG</a>
             <article class="short-f-note">
-                <h2>2015 Creative Commons Korea Some Rights Reserved.</h2>
+                <h2>2023 Creative Commons Korea Some Rights Reserved.</h2>
                 <p>
                     Lorem ipsum, dolor sit amet consectetur adipisicing elit. Modi.
                 </p>
