+++ pages/about.html
@@ -55,7 +55,7 @@
             <section class="headline-abt">
                 <h2 class="grtings mb">"Hello Sharing world"</h2>
                 <h2 class="title-abt">
-                    CREATIVE COMMONS<br>GLOBAL SUMMIT 2015
+                    CREATIVE COMMONS<br>GLOBAL SUMMIT 2023
                 </h2>
                 <p class="intro-txt-abt">
                     The Global Summit brings together the community of experts, academics, and activists who comprise the Creative Commons affiliate network in a different country every two years.
@@ -63,7 +63,7 @@
                     So if you're active and engaged in the worlds of open content and knowledge - free software advocates, wikipedians, Open knowledge, galleries, libraries, museums, archives, governments and foundations, lawyers, and activists - we hope you'll consider joining us this year to build a stronger, more vibrant commons together
                 </p>
                 <p class="intro-txt-abt-1">
-                    Please contact us per Email for nay further questions about CC Global Summit 2015!
+                    Please contact us per Email for nay further questions about CC Global Summit 2023!
                 </p>
                 <a class="email intro-txt-abt-1" href="mailto:sagiecyber@gmail.com">sagiecyber@gmail.com</a>
             </section>
@@ -72,10 +72,10 @@
     <main>
         <section class="main-abt">
             <h2 class="title-abt-1">
-                Creative Commons Global Summit 2015 Logo
+                Creative Commons Global Summit 2023 Logo
             </h2>
             <p class="intro-txt-abt-2">
-                The logo of CC Global Summit 2015 was decided through the logo competition from 8, June to 7, July. 
+                The logo of CC Global Summit 2023 was decided through the logo competition from 8, June to 7, July. 
                 After online public voting and Selection Committee’s judging process, the logo from Naresh Agrawal won the prize. 
                 After consultation with Mr. Agrawal, we optimized the prize-winning-logo for this year’s summit as follows:
             </p>
@@ -114,7 +114,7 @@
         <section class="buttom-f-nav-abt">
             <a class="logo-abt-f" href="#">BLK-NG</a>
             <article class="short-f-note">
-                <h2>2015 Creative Commons Korea Some Rights Reserved.</h2>
+                <h2>2023 Creative Commons Korea Some Rights Reserved.</h2>
                 <p>
                     Lorem ipsum, dolor sit amet consectetur adipisicing elit. Modi.
                 </p>
